pyproject.toml
[project]
name = "ic-hack-26"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "lettucedetect[api]>=0.1.8",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.3",
    "uvicorn>=0.40.0",
]
